+++ runzi/automation/schedule_tasks.py
@@ -10,7 +10,8 @@ from typing import Any, Optional
 
 import boto3
 
-from runzi.automation.local_config import CLUSTER_MODE, WORKER_POOL_SIZE, EnvMode
+from runzi.automation import local_config
+from runzi.automation.local_config import WORKER_POOL_SIZE, ClusterModeEnum
 
 
 def schedule_tasks(scripts: Sequence[Any], worker_pool_size: Optional[int] = None):
@@ -21,30 +22,25 @@ def schedule_tasks(scripts: Sequence[Any], worker_pool_size: Optional[int] = Non
     def call_script(script_name):
         print("call_script with:", script_name)
         try:
-            if CLUSTER_MODE:
+            if local_config.CLUSTER_MODE is ClusterModeEnum.CLUSTER:
                 check_call(['qsub', script_name])
             else:
                 check_call(['bash', script_name])
         except Exception as err:
             print(f"check_call err: {err}")
 
-    if CLUSTER_MODE == EnvMode['LOCAL']:
+    if local_config.CLUSTER_MODE is ClusterModeEnum.AWS:
+        batch_client = boto3.client(
+            service_name='batch', region_name='us-east-1', endpoint_url='https://batch.us-east-1.amazonaws.com'
+        )
+        for script_or_config in scripts:
+            print('AWS_CONFIG: ', script_or_config)
+            res = batch_client.submit_job(**script_or_config)
+            print(res)
+    else:
         print('task count: ', len(scripts))
         print('worker count: ', worker_pool_size)
         pool = Pool(worker_pool_size)
         pool.map(call_script, scripts)
         pool.close()
         pool.join()
-
-    elif CLUSTER_MODE == EnvMode['AWS']:
-
-        batch_client = boto3.client(
-            service_name='batch', region_name='us-east-1', endpoint_url='https://batch.us-east-1.amazonaws.com'
-        )
-        # region_name='ap-southeast-2',
-        # endpoint_url='https://batch.ap-southeast-2.amazonaws.com')
-
-        for script_or_config in scripts:
-            print('AWS_CONFIG: ', script_or_config)
-            res = batch_client.submit_job(**script_or_config)
-            print(res)